Fixes #63 separate scrapers are built for each search engine
