+++ .gitignore
@@ -3,3 +3,7 @@ __pycache__/*
 node_modules/*
 .coverage
 .idea/*
+*.pyc
+*.py.bak
+*.swp
+app/static/bower_components/
